Observer 작동방식 변경
