--- app/src/main/java/com/webrtc/boyj/api/peer/DefaultSdpObserver.java
@@ -1,22 +1,0 @@
-package com.webrtc.boyj.api.peer;
-
-import org.webrtc.SdpObserver;
-import org.webrtc.SessionDescription;
-
-public abstract class DefaultSdpObserver implements SdpObserver {
-
-    @Override
-    public abstract void onCreateSuccess(SessionDescription sessionDescription);
-
-    @Override
-    public void onSetSuccess() {
-    }
-
-    @Override
-    public void onCreateFailure(String s) {
-    }
-
-    @Override
-    public void onSetFailure(String s) {
-    }
-}
