+++ app/src/main/java/com/webrtc/boyj/api/peer/PeerConnectionClient.java
@@ -4,11 +4,14 @@ import android.support.annotation.NonNull;
 
 import com.webrtc.boyj.api.peer.manager.PeerConnectionFactoryManager;
 
+import org.webrtc.DataChannel;
 import org.webrtc.IceCandidate;
 import org.webrtc.MediaConstraints;
 import org.webrtc.MediaStream;
 import org.webrtc.PeerConnection;
 import org.webrtc.PeerConnectionFactory;
+import org.webrtc.RtpReceiver;
+import org.webrtc.SdpObserver;
 import org.webrtc.SessionDescription;
 
 import java.util.ArrayList;
@@ -52,7 +55,7 @@ public class PeerConnectionClient {
     }
 
     @NonNull
-    private BoyjPeerConnection boyjPeerConnection;
+    private PeerConnection peerConnection;
     @NonNull
     private PublishSubject<SessionDescription> sdpSubject = PublishSubject.create();
     @NonNull
@@ -61,11 +64,10 @@ public class PeerConnectionClient {
     private PublishSubject<MediaStream> remoteMediaStreamSubject = PublishSubject.create();
 
     public void createPeerConnection(boolean isCaller) {
-        final PeerConnection peerConnection = peerConnectionFactory.createPeerConnection(rtcConfiguration, new PeerConnectionObserver());
-        boyjPeerConnection = new BoyjPeerConnection(peerConnection);
+        final PeerConnection peerConnection = peerConnectionFactory.createPeerConnection(rtcConfiguration, new BoyjPeerConnectionObserver());
 
         if (isCaller) {
-            boyjPeerConnection.createOffer();
+            peerConnection.createOffer(new BoyjSdpObserver(), constraints);
         }
     }
 
@@ -74,7 +76,7 @@ public class PeerConnectionClient {
     }
 
     public void release() {
-        boyjPeerConnection.dispose();
+        peerConnection.dispose();
         peerConnectionFactory.dispose();
     }
 
@@ -91,52 +93,87 @@ public class PeerConnectionClient {
     }
 
 
-    private class BoyjPeerConnection {
-        @NonNull
-        final private PeerConnection peer;
+    private class BoyjPeerConnectionObserver implements PeerConnection.Observer {
 
-        public BoyjPeerConnection(PeerConnection peer) {
-            this.peer = peer;
+        @Override
+        public void onSignalingChange(PeerConnection.SignalingState signalingState) {
+
         }
 
-        private void createOffer() {
-            peer.createOffer(new LocalSdpObserver(), constraints);
+        @Override
+        public void onIceConnectionChange(PeerConnection.IceConnectionState iceConnectionState) {
+
         }
 
-        private void createAnswer() {
-            peer.createAnswer(new LocalSdpObserver(), constraints);
+        @Override
+        public void onIceConnectionReceivingChange(boolean b) {
+
         }
 
-        public void setRemoteDescription(SessionDescription sdp) {
-            peer.setRemoteDescription(null, sdp);
+        @Override
+        public void onIceGatheringChange(PeerConnection.IceGatheringState iceGatheringState) {
+
         }
 
-        public void addIceCandidate(IceCandidate iceCandidate) {
-            peer.addIceCandidate(iceCandidate);
-        }
-
-        public void dispose() {
-            peer.dispose();
-        }
-    }
-
-    private class PeerConnectionObserver extends DefaultPeerConnectionObserver {
         @Override
         public void onIceCandidate(IceCandidate iceCandidate) {
             iceCandidateSubject.onNext(iceCandidate);
         }
 
         @Override
+        public void onIceCandidatesRemoved(IceCandidate[] iceCandidates) {
+
+        }
+
+        @Override
         public void onAddStream(MediaStream mediaStream) {
             remoteMediaStreamSubject.onNext(mediaStream);
         }
-    }
 
-    private class LocalSdpObserver extends DefaultSdpObserver {
         @Override
-        public void onCreateSuccess(SessionDescription sessionDescription) {
-            boyjPeerConnection.setRemoteDescription(sessionDescription);
-            sdpSubject.onNext(sessionDescription);
+        public void onRemoveStream(MediaStream mediaStream) {
+
+        }
+
+        @Override
+        public void onDataChannel(DataChannel dataChannel) {
+
+        }
+
+        @Override
+        public void onRenegotiationNeeded() {
+
+        }
+
+        @Override
+        public void onAddTrack(RtpReceiver rtpReceiver, MediaStream[] mediaStreams) {
+
         }
     }
+
+    private class BoyjSdpObserver implements SdpObserver {
+
+        @Override
+        public void onCreateSuccess(SessionDescription sessionDescription) {
+            peerConnection.setRemoteDescription(null, sessionDescription);
+            sdpSubject.onNext(sessionDescription);
+        }
+
+        @Override
+        public void onSetSuccess() {
+
+        }
+
+        @Override
+        public void onCreateFailure(String s) {
+
+        }
+
+        @Override
+        public void onSetFailure(String s) {
+
+        }
+    }
+
+
 }
